Api 함수 모듈화하기 - 데이터 성격 끼리 묶기
기본 인스턴스와 특정 정보가 필요한 인스턴스 구분하기
api 함수 모듈화하기
