+++ src/api/index.js
@@ -1,34 +1,24 @@
 import axios from 'axios';
 import { setInterceptors } from './common/interceptors';
 
-// 액시오스 초기화 함수
+// 기본 인스턴스 : 인증 토큰이 없는 API 요청에 사용한다.
 function createInstance() {
+  return axios.create({ baseURL: process.env.VUE_APP_API_URL });
+}
+
+// 인증 토큰을 포함한 인스턴스
+// url : posts 전용 인스턴스 생성하기 위한 인자
+function createInstanceWithAuth(url) {
   const instance = axios.create({
-    baseURL: process.env.VUE_APP_API_URL,
+    baseURL: `${process.env.VUE_APP_API_URL}${url}`,
   });
   return setInterceptors(instance);
 }
-const instance = createInstance();
 
-// 회원가입 API
-function registerUser(userData) {
-  // axios.post('localhost:8080/signup', userData);
-  return instance.post('signup', userData);
-}
+export const instance = createInstance();
+export const posts = createInstanceWithAuth('posts');
 
-// 로그인 API
-function loginUser(userData) {
-  return instance.post('login', userData);
-}
-
-// 학습 노트 데이터를 조회 API
-function fetchPosts() {
-  return instance.get('posts');
-}
-
-// 학습 노트 데이터 생성 API
-function createPost(postData) {
-  return instance.post('posts', postData);
-}
-
-export { registerUser, loginUser, fetchPosts, createPost };
+// get - posts
+// post - posts
+// put - posts {id}
+// delete - posts {id}
