+++ src/components/SignupForm.vue
@@ -22,7 +22,7 @@
 </template>
 
 <script>
-import { registerUser } from '@/api/index';
+import { registerUser } from '@/api/auth';
 import { validateEmail } from '@/utils/validation';
 
 export default {
