+++ src/components/posts/PostAddForm.vue
@@ -37,7 +37,7 @@
 </template>
 
 <script>
-import { createPost } from '@/api/index';
+import { createPost } from '@/api/posts';
 
 export default {
   data() {
