+++ src/views/MainPage.vue
@@ -18,9 +18,9 @@
 </template>
 
 <script>
-import PostListItem from '@/components/posts/PostListItem.vue';
+import PostListItem from '@/components/posts/postListItem.vue';
 import LoadingSpinner from '@/components/common/LoadingSpinner.vue';
-import { fetchPosts } from '@/api/index';
+import { fetchPosts } from '@/api/posts';
 
 export default {
   components: {
